Import lbh-frontend
lbh-frontend has govuk-frontend as a peer dependency.

We have decided to use lbh-frontend directly, to pull in the
stylesheets, rather than using govuk-react. See the ADR for more
context on this decision.

We include them as development dependencies, as there's generally no
need for a consumer of this library to also need the full dependencies,
so we can reduce the bundle size by rolling up the bits we _do_ need.