Add ErrorMessage component
Added ErrrorMessage component for when an error displays in a form. We
have deviated from the govuk-frontend component with regards to
displaying `text` or `html`. It's more idiomatic React to pass in
children, even if it's only a text node being passed in.

Deleted Page component

Page component was an example component, will not be getting used,
and thus will be removed.

Installed classnames package

We need to have dynamic class names in the case of optional classes
being added to components.

Installed prop-types package

We need to type props outside of TypeScript, prop-types allows us to
do that.