Create a pull request template
We want to ensure PRs come in with sufficient context to reviewers. We
also want to use Jira's smart commits feature, to connect work to
tickets. Doing this in feature branch commits causes a lot of duplicate
comments / actions in Jira when rebasing, so putting smart commit
commands in pull requests means those updates would happen on merge.