Modify builds to build a full hierarchy for both CJS and ESM
We want to be able to offer the benefits of being able to import
individual components as needed, to reduce bundle sizes, to all
consumers of this package. Making this change supports that, while
keeping the dependency small by only including type declarations where
they will be used (by ESM imports).