Make sure build runs while testing
Previously, when you run `npm run test`, the project was not rebuilding.
This led to issues where the CSS compilation was not happening, and build
errors were not showing up locally. This fix ensures that `npm run build`
runs everytime `npm run test` and `npm run test:update` are ran.