Make Rollup error on warnings
There are some fairly important warnings that could be considered
errors, one being that stylesheet imports that had an incorrect
path were being flagged as warnings. We want to make sure these
do not go unnnoticed, so Rollup will need to error when it sees
warnings. Note: It will display all warnings before it throws
an error. Currently, some empty chunks will be generated. We don't
want these to break the build, so there is conditional logic to not
throw an error if these are the only warnings.