Remove GOVUK javascript from buttons
GOVUK currently polyfills document within their javascript code for
buttons. We may wish to run these components with a server, which
means that document would not be available. Therefore, we would
either have to polyfill document in other applications or remove
the javascript. In this case, we have decided to remove the javascript
from GOVUK, so we do not need to polyfill within server side applications
using the components, at the cost of having to re-write the javascript
implementation.