Create a Header component
This is based on the existing implementation in `lbh-frontend`.
It's missing a PNG image fallback, as it's not in spec, and we don't
know where the user will be serving them from. Asking them to download
an image separately, or have a build system for importing images feels
like a bad solution.

The tests write to a snapshot file instead of inline, because the markup
is very long.